Fix crashing bug with reels that have only one type of value
When, say, all 10 symbols in one of the reels were hearts, and the machine was predetermined to succeed, then there was a chance _fillPredeterminedValues would assign the 10-heart reel to fail (because at the start of this method, all reels were set to fail).

The fix also involves setting up only the minimum number of reels to hit a number. The rest is left to null, which means that the reel’s result will be random and not pre-determined. This allows for much more believable results as well.