Implement reroll functionality
Reroll is an optional functionality in which the user can reroll failed reels by spending some resource. The resource is completely transparent to the slot machine. It will only see whether or not reroll is enabled, and if so, what to present to the user.

Bump to 4.0 because a bunch of stuff has been deprecated.