Add computation for a floating point wheel
Instead of slot reels (with a finite amount of Xs and Os), we are now using “wheels of fortune”. These can have any probability of success. We use them in a row of five, and we want to know what the probability is of them landing on a success with at least a 3:2 advantage.